backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
email-validator==2.1.0
PyJWT==2.8.0
